apps/compat: revert float64 workaround to idiomatic int types; comparator now normalizes "integer" ↔ "number"
The float64-everywhere choice in the dynamic batch was the wrong direction
— it let TypeScript's "every number is a 64-bit float" worldview
constrain mcpkit's richer Go type system. JSON Schema "integer" is a
subtype of "number"; any integer value validates against both. mcpkit's
"integer" emission is the more precise description.

Real fix: normalize "integer" → "number" in the drift comparator. Now
fixtures use Go-idiomatic types (int for counts and byte totals, float64
for ratios and percentages). Both sides keep their natural emit
behavior; the comparator treats them as equivalent.

Affected:

- system-monitor: CPU count, byte totals, percent (whole), CPU ticks,
  uptime seconds → int / int64 / uint64
- cohort-heatmap: counts and indexes → int; Retention stays float64
  (ratio)
- customer-segmentation: revenue, employee count, account age, support
  tickets, NPS, count → int; EngagementScore stays float64
- debug-server: counter, delayMs, large input length → int / int64
- integration: unchanged (no numeric fields)

Issue 548's Gap 1 entry is being downgraded — this isn't a mcpkit
schema-gen gap, it's a comparator concern, now resolved.

+++ examples/apps/compat/customer-segmentation/main.go
@@ -2,10 +2,11 @@
 //
 // Single get-customer-data tool with an optional enum input ("All" + the four
 // SEGMENTS) and a structured output of customers + segment summaries. No
-// commas in the input description, so struct tags work. Numerics are float64
-// throughout to match upstream's zod-from-`z.number()` "type": "number".
-// Visual test is a random scatter plot upstream masks at the host level —
-// our stub returns empty arrays; the iframe generates its own demo data.
+// commas in the input description, so struct tags work. Idiomatic Go types:
+// counts and dollar/score figures are int; EngagementScore stays float64
+// (it's typically a 0-100 fractional). Visual test is a random scatter
+// plot upstream masks at the host level — our stub returns empty arrays;
+// the iframe generates its own demo data.
 //
 // Run:  EXT_APPS_DIR=/tmp/ext-apps PORT=3101 go run .
 package main
@@ -31,18 +32,18 @@ type customer struct {
 	ID              string  `json:"id"`
 	Name            string  `json:"name"`
 	Segment         string  `json:"segment"`
-	AnnualRevenue   float64 `json:"annualRevenue"`
-	EmployeeCount   float64 `json:"employeeCount"`
-	AccountAge      float64 `json:"accountAge"`
+	AnnualRevenue   int     `json:"annualRevenue"`
+	EmployeeCount   int     `json:"employeeCount"`
+	AccountAge      int     `json:"accountAge"`
 	EngagementScore float64 `json:"engagementScore"`
-	SupportTickets  float64 `json:"supportTickets"`
-	NPS             float64 `json:"nps"`
+	SupportTickets  int     `json:"supportTickets"`
+	NPS             int     `json:"nps"`
 }
 
 type segmentSummary struct {
-	Name  string  `json:"name"`
-	Count float64 `json:"count"`
-	Color string  `json:"color"`
+	Name  string `json:"name"`
+	Count int    `json:"count"`
+	Color string `json:"color"`
 }
 
 type customerDataOutput struct {
